Fix warnings about long long versus the implementation-specific definition of uint64_t. From what I can tell, values of type uint64_t are not meant to be used with printf.
git-svn-id: https://smlnj-gforge.cs.uchicago.edu/svn/manticore/branches/swp@4694 ac4fbcad-d21d-0410-b760-c7b1d3380187

+++ src/lib/parallel-rt/config/gen-basis-offsets.c
@@ -21,7 +21,7 @@ static int LogBufSz = LOGBUF_SZ;
 
 /* print the definition of a symbol */
 #define PR_DEFINE(symb, val)							\
-	printf("#define " #symb " %#0llx\n", (uint64_t)val)
+	printf("#define " #symb " %#0llx\n", (unsigned long long int)val)
 
 /* print a value definition and record it in the CRC buffer */
 #define PR_VALUE(tag, var, value)						\
